Update README.md
Fixed some typos and wording for the change directory step. Changed the hard coded version in the example command from "0.16.0" to "*" so users don't get an error looking for a version they likely haven't actually downloaded. And removed collectors "cs" and "os" as they are deprecated and will be removed entirely this year (see PR below):

https://github.com/prometheus-community/windows_exporter/pull/1596